+++ src/OpenTelemetry.Instrumentation.Process/CHANGELOG.md
@@ -9,6 +9,9 @@
     (its unit changed from `{threads}` to `{thread}`).
   ([#1643](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1643))
 
+* Update `OpenTelemetry.Api` to `1.8.1`.
+  ([#1668](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/1668))
+
 ## 0.5.0-beta.5
 
 Released 2024-Apr-05